# docs/swagger.yaml
basePath: /
definitions:
  dto.ConfigResponse:
    properties:
      Port:
        example: "8080"
        type: string
      Secret:
        example: ok
        type: string
      Status:
        example: ok
        type: string
      http://localhost:27017:
        example: ok
        type: string
    required:
    - Port
    - Secret
    - Status
    - http://localhost:27017
    type: object
  dto.CreatePaymentRequest:
    properties:
      Amount:
        example: 1500
        type: integer
      CategoryID:
        example: cat_123
        type: string
    required:
    - Amount
    - CategoryID
    type: object
  dto.ErrorResponse:
    properties:
      error:
        example: internal server error
        type: string
    required:
    - error
    type: object
  dto.HealthResponse:
    properties:
      Status:
        example: ok
        type: string
    required:
    - Status
    type: object
  dto.LoginRequest:
    properties:
      Email:
        type: string
      Password:
        type: string
    required:
    - Email
    - Password
    type: object
  dto.LoginResponse:
    properties:
      Token:
        type: string
    required:
    - Token
    type: object
  dto.PaymentResponse:
    properties:
      Amount:
        example: 1500
        type: integer
      CategoryID:
        example: cat_123
        type: string
      CreateDate:
        example: "2026-01-20T00:00:00Z"
        type: string
      ID:
        example: 64f8c...
        type: string
      UpdateDate:
        example: "2026-01-20T00:00:00Z"
        type: string
    required:
    - Amount
    - CategoryID
    - CreateDate
    - ID
    - UpdateDate
    type: object
  dto.RegisterRequest:
    properties:
      Email:
        example: test@mail.com
        type: string
      Password:
        example: "123456"
        type: string
    required:
    - Email
    - Password
    type: object
  dto.UserResponse:
    properties:
      CreatedDate:
        example: "2026-01-20T00:00:00Z"
        type: string
      Email:
        example: test@mail.com
        type: string
      ID:
        example: 64f8c...
        type: string
    required:
    - CreatedDate
    - Email
    - ID
    type: object
info:
  contact: {}
  description: Personal finance ledger API
  title: Ledger API
  version: "1.0"
paths:
  /auth/create:
    post:
      consumes:
      - application/json
      description: Create a new user with email and password
      parameters:
      - description: User registration data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Register new user
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Login
      tags:
      - auth
  /auth/protected:
    get:
      description: Check cookies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HealthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Protected
      tags:
      - auth
  /config:
    get:
      description: Check service config
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ConfigResponse'
      summary: Get config
      tags:
      - system
  /health:
    get:
      description: Check service availability
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HealthResponse'
      summary: Health check
      tags:
      - system
  /payments:
    get:
      description: Get a list of all payments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PaymentResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all payments
      tags:
      - payments
    post:
      consumes:
      - application/json
      description: Create a payment with amount and category
      parameters:
      - description: Payment data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Create a new payment
      tags:
      - payments
  /users:
    get:
      description: Get a list of all users (emails only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"